django-backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
django-resized==1.0.2
djangorestframework==3.15.1
gunicorn==21.2.0
Markdown==3.6
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.11.0

